report hourly outputs more efficiently
